+++ packages/@react-native-windows/bot-coordinator/package.json
@@ -13,12 +13,14 @@
     "watch": "just-scripts watch"
   },
   "dependencies": {
-    "@react-native-windows/bot-actors": "1.0.0"
+    "@react-native-windows/bot-actors": "1.0.0",
+    "lodash": "^4.17.15"
   },
   "devDependencies": {
     "@azure/functions": "^1.0.2-beta2",
     "@rnw-scripts/just-task": "0.0.6",
     "@rnw-scripts/ts-config": "0.1.0",
+    "@types/lodash": "^4.14.161",
     "eslint": "7.12.0",
     "just-scripts": "^0.44.7",
     "prettier": "1.19.1",
